move solution to src folder
