+++ MrEric.nuspec
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
+    <metadata>
+        <id>Sizikov.MrEric</id>
+        <version>@build.number@</version>
+        <title>Mr. Eric (for R# 2016)</title>
+        <authors>Anton Sizikov</authors>
+	<iconUrl>https://raw.githubusercontent.com/asizikov/mr-eric/master/Content/lock.png</iconUrl>
+        <owners>AntonSizikov</owners>
+        <projectUrl>https://github.com/asizikov/mr-eric</projectUrl>
+        <requireLicenseAcceptance>false</requireLicenseAcceptance>
+        <description>"Create and initialize private auto-property" context action, quick fix and live template for ReSharper.</description>
+        <releaseNotes>ReSharper 2016.1 support</releaseNotes>
+        <language>en-US</language>
+        <tags>auto-property, autoproperty</tags>
+        <dependencies>
+            <dependency id="Wave" version="[5.0]" />
+        </dependencies>
+    </metadata>
+    <files>
+        <file src=".build\MrEric.R91.dll" target="dotFiles" />
+        <file src="Templates\mr-eric-templates.DotSettings" target="DotFiles\Extensions\Sizikov.MrEric\settings\" />
+    </files>
+</package>
