--- MrEric/MrEric.sln
@@ -1,26 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{EA68B75B-FD4D-4148-BB11-17392638A1E6}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\packages.config = .nuget\packages.config
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MrEric.R91", "MrEric.R91\MrEric.R91.csproj", "{7446AAE8-D6CE-4EA5-B91F-BEE927D378AC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7446AAE8-D6CE-4EA5-B91F-BEE927D378AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7446AAE8-D6CE-4EA5-B91F-BEE927D378AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7446AAE8-D6CE-4EA5-B91F-BEE927D378AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7446AAE8-D6CE-4EA5-B91F-BEE927D378AC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
