+++ src/MrEric.R91/Properties/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("MrEric")]
+[assembly: AssemblyDescription("Create and initialize private auto-property action.")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Anton Sizikov")]
+[assembly: AssemblyProduct("MrEric")]
+[assembly: AssemblyCopyright("Copyright © Anton Sizikov, 2015")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: AssemblyVersion("1.0.5.0")]
+[assembly: AssemblyFileVersion("1.0.5.0")]
