+++ src/MrEric.R91/Psi/PropertyDeclarationExtensions.cs
