+++ src/MrEric.R91/ZoneMarker.cs
@@ -1,0 +1,11 @@
+﻿using JetBrains.Application.BuildScript.Application.Zones;
+using JetBrains.ReSharper.Feature.Services;
+using JetBrains.ReSharper.Psi.CSharp;
+
+namespace MrEric
+{
+    [ZoneMarker]
+    public class ZoneMarker : IRequire<ICodeEditingZone>, IRequire<ILanguageCSharpZone>
+    {
+    }
+}
