+++ src/MrEric.R91/packages.config
@@ -1,0 +1,57 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Antlr2.Runtime" version="2.7.7.02" targetFramework="net40" />
+  <package id="Antlr2.Tools" version="2.7.6.4" targetFramework="net40" />
+  <package id="AsyncBridge.JetBrains" version="0.1.1.0" targetFramework="net40" />
+  <package id="DotNetZip.Reduced" version="1.9.1.8" targetFramework="net40" />
+  <package id="JetBrains.Annotations" version="10.1.4" targetFramework="net40" />
+  <package id="JetBrains.Build.Platform.Tasks.ThemedIconsPacker" version="2.0.20151217.1" targetFramework="net40" developmentDependency="true" />
+  <package id="JetBrains.ExternalAnnotations" version="10.1.13" targetFramework="net40" />
+  <package id="JetBrains.Mono.Posix" version="0.1" targetFramework="net4" />
+  <package id="JetBrains.Platform.Core.Ide" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.Core.Shell" version="105.0.20160520.92915" targetFramework="net40" />
+  <package id="JetBrains.Platform.Core.Text" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.Installer" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.Interop.dotMemoryUnit.Interop" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.DevExpress" version="2.0.20150224.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.Microsoft.Deployment.Compression.Cab" version="2.0.20140304.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.Microsoft.Deployment.WindowsInstaller" version="2.0.20140821.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.System.Windows.Interactivity" version="2.0.20140318.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.VisualStudio.AnyVs.ShellInterop.PrivateBuild" version="2.0.20140304.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Lib.WpfContrib" version="2.0.20150225.0" targetFramework="net40" />
+  <package id="JetBrains.Platform.Symbols" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.Tests.Framework" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Platform.VisualStudio" version="105.0.20160522.80211" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.Core" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.Cpp.Src" version="105.0.20160523.141910" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.SolutionBuilder" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.src" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.Tasks" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.test.Framework" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.UnitTesting" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.Psi.Features.VisualStudio" version="105.0.20160523.140224" targetFramework="net40" />
+  <package id="JetBrains.ReSharper.SDK" version="2016.1.20160523.144409" targetFramework="net40" developmentDependency="true" />
+  <package id="JetBrains.ReSharper.SDK.Internal" version="105.0.20160523.144409" targetFramework="net40" />
+  <package id="JetBrains.Unix.Utils" version="20150702.4" targetFramework="net40" />
+  <package id="JetBrains.WinDbg.SymStore" version="10.0.10586.20151201" targetFramework="net40" developmentDependency="true" />
+  <package id="Microsoft.VC100.CRT.JetBrains" version="10.00.40219.2" targetFramework="net40" />
+  <package id="Microsoft.VC120.CRT.JetBrains" version="12.0.21005.2" targetFramework="net40" />
+  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net40" />
+  <package id="Newtonsoft35.Json" version="7.0.1" targetFramework="net40" />
+  <package id="NuGet.Core" version="2.10.1" targetFramework="net40" />
+  <package id="NUnit" version="2.6.4" targetFramework="net40" />
+  <package id="NUnit.Core" version="2.6.4" targetFramework="net40" />
+  <package id="NUnit.ReSharperRunner2" version="2.6.408" targetFramework="net4" />
+  <package id="NUnit.ReSharperRunner3" version="3.0.12" targetFramework="net40" />
+  <package id="NVelocity" version="1.0.3" targetFramework="net40" />
+  <package id="SharpZipLib.JetBrains.Stripped" version="0.86.20150204.1" targetFramework="net40" />
+  <package id="Sprache.JetBrains" version="2.0.0.44" targetFramework="net40" />
+  <package id="Vestris.ResourceLib.JetBrains" version="1.4.20150303.0" targetFramework="net40" />
+  <package id="Wave" version="5.0.0.0" targetFramework="net40" />
+  <package id="WinDbg.SymStore" version="8.1.20130822.1833" targetFramework="net40" developmentDependency="true" />
+  <package id="Windows7APICodePack.JetBrains.Stripped" version="1.1.20150225.0" targetFramework="net40" />
+  <package id="xmlrpcnet" version="2.5.0" targetFramework="net40" />
+  <package id="xunit.abstractions.JetBrains" version="2.0.0" targetFramework="net40" />
+  <package id="xunit.JetBrains" version="1.9.2" targetFramework="net40" />
+  <package id="xunit.runner.utility.JetBrains" version="2.2.0" targetFramework="net40" />
+</packages>
